# Speed up use of PBKDF2 when building in Debug configuration