--- a/Sources/_MongoKittenCrypto/SHA1.swift
+++ b/Sources/_MongoKittenCrypto/SHA1.swift
@@ -17,7 +17,7 @@ public struct SHA1 : Hash {
     
     var f: UInt32 = 0
     var k: UInt32 = 0
-    var temp: UInt32 = 0
+//    var temp: UInt32 = 0
     
     public var processedBytes: UInt64 = 0
     
@@ -57,19 +57,20 @@ public struct SHA1 : Hash {
     }
     
     public mutating func update(from pointer: UnsafePointer<UInt8>) {
-        var w = pointer.withMemoryRebound(to: UInt32.self, capacity: 16, { pointer in
-            return [
-                pointer[0].bigEndian, pointer[1].bigEndian, pointer[2].bigEndian, pointer[3].bigEndian,
-                pointer[4].bigEndian, pointer[5].bigEndian, pointer[6].bigEndian, pointer[7].bigEndian,
-                pointer[8].bigEndian, pointer[9].bigEndian, pointer[10].bigEndian, pointer[11].bigEndian,
-                pointer[12].bigEndian, pointer[13].bigEndian, pointer[14].bigEndian, pointer[15].bigEndian,
-                ]
-        })
-        
-        w.reserveCapacity(80)
-        
-        for i in 16...79 {
-            w.append(leftRotate(w[i &- 3] ^ w[i &- 8] ^ w[i &- 14] ^ w[i &- 16], count: 1))
+        var i = 16
+        let w = [UInt32](unsafeUninitializedCapacity: 80) { buf, count in
+            buf.initialize(repeating: 0)
+            pointer.withMemoryRebound(to: UInt32.self, capacity: 16, { pointer in
+                (buf[ 0], buf[ 1], buf[ 2], buf[ 3]) = (pointer[ 0].bigEndian, pointer[ 1].bigEndian, pointer[ 2].bigEndian, pointer[ 3].bigEndian)
+                (buf[ 4], buf[ 5], buf[ 6], buf[ 7]) = (pointer[ 4].bigEndian, pointer[ 5].bigEndian, pointer[ 6].bigEndian, pointer[ 7].bigEndian)
+                (buf[ 8], buf[ 9], buf[10], buf[11]) = (pointer[ 8].bigEndian, pointer[ 9].bigEndian, pointer[10].bigEndian, pointer[11].bigEndian)
+                (buf[12], buf[13], buf[14], buf[15]) = (pointer[12].bigEndian, pointer[13].bigEndian, pointer[14].bigEndian, pointer[15].bigEndian)
+            })
+            while i < 80 {
+                buf[i] = leftRotate(buf[i &- 3] ^ buf[i &- 8] ^ buf[i &- 14] ^ buf[i &- 16], count: 1)
+                i &+= 1
+            }
+            count = buf.count
         }
         
         a = h0
@@ -78,38 +79,43 @@ public struct SHA1 : Hash {
         d = h3
         e = h4
         
-        for i in 0...79 {
-            switch i {
-            case 0...19:
-                f = (b & c) | ((~b) & d)
-                k = 0x5A827999
-            case 20...39:
-                f = b ^ c ^ d
-                k = 0x6ED9EBA1
-            case 40...59:
-                f = (b & c) | (b & d) | (c & d)
-                k = 0x8F1BBCDC
-            default:
-                f = b ^ c ^ d
-                k = 0xCA62C1D6
+        w.withUnsafeBufferPointer { w in
+            i = 0
+            while i < 80 {
+                let f: UInt32, k: UInt32
+                if i < 20 {
+                    f = (b & c) | ((~b) & d)
+                    k = 0x5A827999
+                } else if i < 40 {
+                    f = b ^ c ^ d
+                    k = 0x6ED9EBA1
+                } else if i < 60 {
+                    f = (b & c) | (b & d) | (c & d)
+                    k = 0x8F1BBCDC
+                } else {
+                    f = b ^ c ^ d
+                    k = 0xCA62C1D6
+                }
+                
+                let temp = leftRotate(a, count: 5) &+ f &+ e &+ w[i] &+ k
+                e = d
+                d = c
+                c = leftRotate(b, count: 30)
+                b = a
+                a = temp
+                i &+= 1
             }
-            
-            temp = leftRotate(a, count: 5) &+ f &+ e &+ w[i] &+ k
-            e = d
-            d = c
-            c = leftRotate(b, count: 30)
-            b = a
-            a = temp
         }
         
-        h0 = h0 &+ a
-        h1 = h1 &+ b
-        h2 = h2 &+ c
-        h3 = h3 &+ d
-        h4 = h4 &+ e
+        h0 &+= a
+        h1 &+= b
+        h2 &+= c
+        h3 &+= d
+        h4 &+= e
     }
 }
 
+@_transparent
 fileprivate func leftRotate(_ x: UInt32, count c: UInt32) -> UInt32 {
-    return (x << c) | (x >> (32 - c))
+    return (x &<< c) | (x &>> (32 &- c))
 }