@gwynne 2023-07-28
For comparison, here's a quick set of performance numbers for using the SwiftCrypto `Insecure.MD5`, `Insecure.SHA1`, and `SHA256` implementations:

#### macOS (Xcode 15 beta 5)
Config|MD5|SHA-1|SHA-256
-:|:-:|-:|:-:
debug|0.141s|0.119s|0.120s
release|0.087s|0.067s|0.067s

#### Linux (swift:5.8-jammy)
Config|MD5|SHA-1|SHA-256
-:|:-:|-:|:-:
debug|0.340s|0.237s|0.258s
release|0.076s|0.063s|0.065s

#### Linux (5.9 nightly - swift-5.9-DEVELOPMENT-SNAPSHOT-2023-07-25-a)
Config|MD5|SHA-1|SHA-256
-:|:-:|-:|:-:
debug|0.352s|0.248s|0.280s
release|0.074s|0.062s|0.064s

Conclusions:
- SwiftCrypto's MD5 implementation is oddly slow. (This might just be a side effect of MD5 having relatively little internal state, or possibly the M1 just handles SHA-1 and 2 better 🤷‍♀️)
- SwiftCrypto's SHA-1 implementation is much faster than us.
- SwiftCrypto's SHA-2 implementation is _massively_ faster than us.

Might be the numbers would change all around on an Intel machine, though.